+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.11] - 202X-XX-XX
+### Changed
+- implemented the $APP_DATA/.initialized file to detect an already installed system
+- docker logs small refactoring
+
 ## [0.9.10] - 2020-09-23
 ### Changed
 - defined release immutable image tag into docker-compose.yml